+++ package.json
@@ -17,6 +17,7 @@
     "react-router-dom": "^6.4.3",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.6",
+    "styled-reset": "^4.4.2",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
