+++ src/pages/BulletinBoard.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function BulletinBoard() {
+  return (
+    <div>BulletinBoard</div>
+  )
+}
